ctf: merge packet context into event content
The (stream-specific) packet context may contain additional
context information which is constant for every event. This is typically
a good place to store large strings. In order to make these available in each
event's content as 'context.*' fields, the packet context must be merged into
the content as it is already done with the stream event context and event
context.